% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_file2tab.R
\name{get_file2tab}
\alias{get_file2tab}
\title{\code{get_file2tab}
Scan txt files and extract information into a data.frame.}
\usage{
get_file2tab(files, features, replace = T)
}
\arguments{
\item{files}{Full names of the files to scan [vector].}

\item{features}{Unique features for grep to search [vector].}
}
\value{
return Extracted values [data.frame].
}
\description{
\code{get_file2tab}
Scan txt files and extract information into a data.frame.
}
\examples{
files <- list.files(path = dir, pattern = fileptn, full.names = TRUE)
features <- c("C methylated in CHH context")
get_file2tab(files, features, replace=T )

}

