% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_mht.R
\name{plot_mht}
\alias{plot_mht}
\title{\code{manhattan plot}}
\usage{
plot_mht(res = res, cex = 0.9, pch = 16, col = rep(c("slateblue",
  "cyan4"), 5), GAP = 5e+06, yaxis = NULL, col2plot = "ModelFreq", ...)
}
\arguments{
\item{inputdf}{An input data.frame, with columns of file and out.}

\item{outdir}{The dir of shell files.}

\item{cmdno}{Number of commands to excute in each array.}

\item{rcodes}{The abosulte path of your R codes to run.}

\item{arrayshid}{The sbatch id.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{cmdno}{Number of commands per CPU, i.e. number of rows per inputdf.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
\code{manhattan plot}
}
\examples{
run_Rcodes(inputdf=data.frame(file=1:11, out=10), outdir="slurm-script", cmdno=10,
           rcodes = "lib/C_format.R", arrayshid = "slurm-script/run_rcode_array.sh",
           email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

