% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_alphaImpute.R
\name{run_alphaimpute}
\alias{run_alphaimpute}
\alias{set_alphaimpute}
\title{\code{Run GATK Joint Genotype on farm}}
\usage{
run_alphaimpute(pedfile = "data/Parentage_for_imputeR.csv", geno,
  outputdir = "largedata/teo_alpha",
  EditingParameters = "95.0,2.0,98.0,AllSnpOut", RestartOption = 1,
  email = NULL, runinfo = c(FALSE, "bigmemh", 4))

set_alphaimpute(ped, chrgeno, numsnp, EditingParameters, RestartOption, runinfo,
  shid)
}
\arguments{
\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{fq}{An input data.frame for fastq files. Must contains fq1, fq2 and out.}

\item{kitpath}{The absolute or relative path of the fermi.kit directory that can invoke the pipeline.}

\item{genome}{The full path of genome with bwa indexed reference fasta file.}

\item{s}{Approximate genome size, default=3g.}

\item{t}{Number of threads, default=16.}

\item{l}{Primary read length, default=100.}

\item{arrayjobs}{A character specify the number of array you try to run, i.e. 1-100.}

\item{jobid}{The job name show up in your sq NAME column.}
}
\value{
return a batch of shell scripts.
}
\description{
GATK Best Practices: recommended workflows for variant discovery analysis.
}
\details{
see more detail about GATK:
\url{https://www.broadinstitute.org/gatk/guide/bp_step.php?p=1}

local programs:
bwa Version: 0.7.5a-r405
picard-tools-2.1.1
GenomeAnalysisTK-3.5/
}
\examples{
fq <- data.frame(fq1="fq_1.fq", fq2="f1_2.fq", out="mysample",
                 group="g1", sample="s1",PL="illumina", LB="lib1", PU="unit1")
run_fermikit(fq, kitpath="/home/jolyang/bin/fermikit/",
genome="/home/jolyang/dbcenter/AGP/AGPv2", s='3g', t=16, l=100, arrayjobs="1-2",
jobid="fermi", email=NULL)

}

