% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_ANGSD.R
\name{run_ANGSD}
\alias{run_ANGSD}
\title{\code{Run ANGSD on farm}}
\usage{
run_ANGSD(type = "GL", shfile = "slurm-script/run_angsd_cmd.sh",
  bamlist = "list.txt", outfile = "out", ref = "ref.fa", nInd = 20,
  minInd = 10, indF = "indF.txt", anc = "ref.fa", email = NULL,
  runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}

\item{inputdf}{An input data.frame.}

\item{minQ}{Minimum base quality score.}

\item{minMapQ}{[int]=0 Minimum mapQ quality.}

\item{cmdno}{Number of commands per CPU, i.e. number of rows per inputdf.}
}
\value{
return a batch of shell scripts.
}
\description{
angsd
}
\details{
see more detail about ANGSD
\url{http://popgen.dk/angsd/index.php/Main_Page}

Scripts learned from Tim's paper \url{https://github.com/timbeissinger/Maize-Teo-Scripts}
}
\examples{
run_ANGSD(type="GL", shfile="slurm-script/run_angsd_cmd.sh",
bamlist="list.txt", outfile="out", ref="ref.fa", nInd=20, minInd=10,
indF="indF.txt", anc="ref.fa",
email=NULL, runinfo = c(FALSE, "bigmemh", 1)


}

