% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_aspera.R
\name{run_aspera}
\alias{run_aspera}
\title{\code{Run Aspera Connect to download from SRA.}}
\usage{
run_aspera(sra, maxspeed = "200m", outdir = ".", arrayjobs = "1-2",
  jobid = "aspera", email = NULL)
}
\arguments{
\item{sra}{An input data.frame for SRA ids. Must contains column: SRR.}

\item{maxspeed}{The max speed for downloading.}

\item{outdir}{The output directory.}

\item{arrayjobs}{A character specify the number of array you try to run, i.e. 1-100.}

\item{jobid}{The job name show up in your sq NAME column.}

\item{email}{Your email address that farm will email to once the job was done/failed.}
}
\value{
return a batch of shell scripts.
}
\description{
Downloading SRA using 'ascp' utility or Aspera Connect.
}
\details{
see more detail about SRA with Aspera downloading:
\url{http://www.ncbi.nlm.nih.gov/books/NBK158899/#SRA_download.downloading_sra_data_using}
}
\examples{
sra <- data.frame(SRR=c("ERR877647", "ERR877648"),
SRX=c( "ERX957210", "ERX957211"),
pid=c( "1_Base1_Bbreve-sc-2188486", "P1_ECvsrS_1-sc-2201977"))
run_aspera(sra, maxspeed="200m", outdir=".", arrayjobs="1-2", jobid="aspera", email=NULL)

}

