% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_bs2VCF.R
\name{run_bs2vcf}
\alias{run_bs2vcf}
\title{\code{Run bs2vcf job on farm}}
\usage{
run_bs2vcf(inputdf, email = NULL, runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{inputdf}{An input data.frame for fastq files. Must contains bsmap and out.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
convert BSMAP format to VCF format
Find help about 'bs2vcf -h'.
bs2vcf -p /group/jrigrp4/BS_teo20/BSMAP_output -i test_methratio.txt -o test_methratio.vcf
}
\examples{
inputdf <- data.frame(bsmap="test_ratio.txt", out="test_ratio.vcf")
run_bs2vcf( inputdf, email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

