% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_cutadapt.R
\name{run_cutadapt}
\alias{run_cutadapt}
\title{\code{Run cutadapt}}
\usage{
run_cutadapt(inputdf, ad1 = "AGATCGGAAGAGC", ad2 = "AGATCGGAAGAGC",
  q = 20, email = NULL, runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{inputdf}{Data.frame with fq1, fq2, out1 and out2 columns. Both input and output can be gzipped (.gz).}

\item{ad1}{Adapter 3' method, see cutadapt documentation for more details. Same for ad2.}

\item{q}{quality_cutoff Default=20. Note:use -q0 to get the distribution of all quality values}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return A batch of shell scripts.
}
\description{
install in .local/bin using 'pip install --user --upgrade cutadapt'.
}
\details{
For more information:
https://cutadapt.readthedocs.io/en/stable/index.html
version: v1.10
}
\examples{
fqs <- c("$HOME/dbcenter/Ecoli/fastq/SRR2921970.sra_1.fastq.gz",
         "$HOME/dbcenter/Ecoli/fastq/SRR2921970.sra_2.fastq.gz")
inputdf <- data.frame(fq1=fqs[1], fq2=fqs[2], out1=fqs[1], out2=fqs[2])
df$out <- gsub("sra_.*", "qc", df$out)
run_fastq_qc(df, email=NULL, runinfo = c(FALSE, "bigmemh", 1))

}

