% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_fq_dump.R
\name{run_fq_dump}
\alias{run_fq_dump}
\title{\code{Run fastq-dump.}}
\usage{
run_fq_dump(filepath = "/group/jrigrp4/BS_teo20/WGBS",
  slurmsh = "slurm-script/dump_WGBS.sh", rmsra = TRUE, email = NULL)
}
\arguments{
\item{filepath}{The absolute path of the SRA files.}

\item{slurmsh}{File name of the output shell command.}

\item{rmsra}{Remove the original SRA file after dumpping.}

\item{email}{Your email address that farm will email to once the job was done/failed.}
}
\value{
return a single shell script to run.
}
\description{
fastq-dump to dump SRA file.
}
\details{
see more detail about SRA with Aspera downloading:
\url{http://www.ncbi.nlm.nih.gov/books/NBK158899/#SRA_download.downloading_sra_data_using}
}
\examples{
run_fq_dump(filepath="/group/jrigrp4/BS_teo20/WGBS",
slurmsh="slurm-script/dump_WGBS.sh",
rmsra=TRUE, email=NULL)

}

