% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GATK.R
\name{run_GATK}
\alias{run_GATK}
\title{\code{Run GATK job on farm}}
\usage{
run_GATK(inputdf,
  ref.fa = "~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
  gatkpwd = "$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
  picardpwd = "$HOME/bin/picard-tools-2.1.1/picard.jar", markDup = TRUE,
  realignInDels = FALSE, indels.vcf = "indels.vcf", recalBases = FALSE,
  dbsnp.vcf = "dbsnp.vcf", email = NULL, runinfo = c(TRUE, "bigmemh", 4))
}
\arguments{
\item{inputdf}{An input data.frame for fastq files. Must contains fq1, fq2, out (and/or bam).
If inputdf contained bam, bwa alignment will be scaped.}

\item{ref.fa}{The full path of genome with bwa indexed reference fasta file.}

\item{gatkpwd}{The absolute path of GenomeAnalysisTK.jar.}

\item{picardpwd}{The absolute path of picard.jar.}

\item{markDup}{Primary read length, default=100.}

\item{realignInDels}{A character specify the number of array you try to run, i.e. 1-100.}

\item{indels.vcf}{The job name show up in your sq NAME column.}

\item{recalBases}{The job name show up in your sq NAME column.}

\item{dbsnp.vcf}{The job name show up in your sq NAME column.}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{runinfo}{It will pass to \code{set_array_job}.}

\item{inputbam}{The job name show up in your sq NAME column.}
}
\value{
return a batch of shell scripts.
}
\description{
GATK Best Practices: recommended workflows for variant discovery analysis.
}
\details{
see more detail about GATK:
\url{https://www.broadinstitute.org/gatk/guide/bp_step.php?p=1}

local programs:
bwa Version: 0.7.5a-r405
picard-tools-2.1.1
GenomeAnalysisTK-3.5/
}
\examples{
inputdf <- data.frame(fq1="fq_1.fq", fq2="f1_2.fq", out="mysample",
                 group="g1", sample="s1", PL="illumina", LB="lib1", PU="unit1")

run_GATK(inputdf, ref.fa="~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
         gatkpwd="$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
         picardpwd="$HOME/bin/picard-tools-2.1.1/picard.jar",
         markDup=TRUE, realignInDels=TRUE, indels.vcf="indels.vcf",
         recalBases=TRUE, dbsnp.vcf="dbsnp.vcf", email=NULL,
         runinfo = c(FALSE, "bigmemh", 4))

}

