% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GATK_JointGenotype.R
\name{run_GATK_JointGenotype}
\alias{run_GATK_JointGenotype}
\alias{set_hardfilter}
\title{\code{Run GATK Joint Genotype on farm}}
\usage{
run_GATK_JointGenotype(gvcf, outvcf,
  ref.fa = "~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
  gatkpwd = "$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
  includeNonVariantSites = FALSE, hardfilter = TRUE,
  snpflt = "\\"QD < 2.0 || FS > 60.0 || MQ < 40.0 || MQRankSum < -12.5 || ReadPosRankSum < -8.0\\"",
  indelflt = "\\"QD < 2.0 || FS > 200.0 || ReadPosRankSum < -20.0\\"",
  email = NULL, runinfo = c(FALSE, "bigmemh", 1))

set_hardfilter(outvcf, gatkpwd, snpflt, indelflt, ref.fa, runinfo, shid)
}
\arguments{
\item{gvcf}{A vector of g.vcf files.}

\item{outvcf}{File name of the output vcf files, default="mysamples.vcf".}

\item{ref.fa}{The full path of genome with bwa indexed reference fasta file.}

\item{gatkpwd}{The absolute path of GenomeAnalysisTK.jar.}

\item{includeNonVariantSites}{Include loci found to be non-variant after genotyping (for GenotypeGVCFs).}

\item{hardfilter}{Whether to filter variants. see detail about how to apply hard filters to a call set.
\url{https://www.broadinstitute.org/gatk/guide/article?id=2806}}

\item{snpflt}{Parameters to apply the filter to the SNP call set.}

\item{indelflt}{Parameters to apply the filter to the Indel call set.}

\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
GATK Best Practices: recommended workflows for variant discovery analysis.
}
\details{
see more detail about GATK:
\url{https://www.broadinstitute.org/gatk/guide/bp_step.php?p=1}

local programs:
bwa Version: 0.7.5a-r405
picard-tools-2.1.1
GenomeAnalysisTK-3.5/
}
\examples{
gvcf <- c("1.vcf", "2.vcf")
outvcf <- "out.vcf"
run_GATK_JointGenotype(
gvcf,
outvcf="mysamples.vcf",
ref.fa="~/dbcenter/Ecoli/reference/Ecoli_k12_MG1655.fasta",
gatkpwd="$HOME/bin/GenomeAnalysisTK-3.5/GenomeAnalysisTK.jar",
includeNonVariantSites=FALSE,
hardfilter=TRUE,
snpflt="\\"QD < 2.0 || FS > 60.0 || MQ < 40.0 || MQRankSum < -12.5 || ReadPosRankSum < -8.0\\"",
indelflt="\\"QD < 2.0 || FS > 200.0 || ReadPosRankSum < -20.0\\"",
email=NULL,
runinfo = c(FALSE, "bigmemh", 1) )

}

