% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_gerpIBD.R
\name{run_gerpIBD}
\alias{run_gerpIBD}
\title{\code{Run gerpIBD on farm}}
\usage{
run_gerpIBD(inputdf, email = NULL, runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{inputdf}{An input data.frame. Type "gerpIBD -h" to understand the meaning of the input parameters.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
gerpIBD: compute sum of GERP score in a given IBD region across genotypes.
}
\details{
see more detail about gerpIBD
\url{https://github.com/yangjl/zmSNPtools}
}
\examples{
traits <- tolower(c("ASI", "DTP", "DTS", "EHT", "GY", "PHT", "TW"))
inputdf <- data.frame(trait=rep(traits, each=44),
   mode=rep(rep(c("add", "dom", "k", "k5"), each=11), times=7 ),
   nperm=rep(0:10, times=28),
   d="largedata/IBD/allsnps_11m_IBD.bed",
   s="largedata/SNP/allsnps_11m.dsf5",
   g="largedata/SNP/geno_b0_cs/gerpv2_b0_cs0.csv",
   f="largedata/snpeff/perse/asi_k.txt",
   out="res")

run_gerpIBD(inputdf[1:10,], email=NULL, runinfo = c(FALSE, "bigmemh", 1) )

}

