% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_icmd.R
\name{run_iput}
\alias{run_iput}
\title{\code{Run iget and iput to get/put data from/to iPlant.}}
\usage{
run_iput(files, jobs1cpu = 1, localpwd = "largedata/gatk_vcf",
  email = NULL, runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{email}{Your email address that farm will email to once the job was done/failed.}

\item{sra}{An input data.frame for SRA ids. Must contains column: SRR.}

\item{maxspeed}{The max speed for downloading.}

\item{outdir}{The output directory.}

\item{arrayjobs}{A character specify the number of array you try to run, i.e. 1-100.}

\item{jobid}{The job name show up in your sq NAME column.}
}
\value{
return a batch of shell scripts.
}
\description{
Downloading SRA using 'ascp' utility or Aspera Connect.
}
\details{
see more detail about SRA with Aspera downloading:
\url{http://www.ncbi.nlm.nih.gov/books/NBK158899/#SRA_download.downloading_sra_data_using}
}
\examples{

}

