% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_Rcodes.R
\name{run_Rcodes}
\alias{run_Rcodes}
\title{\code{Run array job of R codes}}
\usage{
run_Rcodes(inputdf, outdir, cmdno = 100, rcodes = "lib/C_format.R",
  arrayshid = "slurm-script/run_bcf_query_array.sh", email = NULL,
  runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{inputdf}{An input data.frame, with columns of file and out.}

\item{outdir}{The dir of shell files.}

\item{cmdno}{Number of commands to excute in each array.}

\item{rcodes}{The abosulte path of your R codes to run.}

\item{arrayshid}{The sbatch id.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}

\item{cmdno}{Number of commands per CPU, i.e. number of rows per inputdf.}
}
\value{
return a batch of shell scripts.
}
\description{
\code{Run array job of R codes}
}
\examples{


}

