% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_arrayjob.R
\name{set_array_job}
\alias{set_array_job}
\title{\code{Set up array job on farm}}
\usage{
set_array_job(shid = "largedata/GenSel/CL_test.sh",
  shcode = "sh largedata/myscript.sh", arrayjobs = "1-700", wd = NULL,
  jobid = "myjob", email = NULL)
}
\arguments{
\item{shid}{Relative or absolute path and file name of your shell code, i.e. CL_test.sh.}

\item{shcode}{The commands inside your sh file.}

\item{arrayjobs}{A character specify the number of array you try to run, i.e. 1-100.}

\item{wd}{Working directory, default=NULL. It will use your current directory.}

\item{jobid}{The job name show up in your sq NAME column.}

\item{email}{Your email address that farm will email to once the job was done or failed.}
}
\value{
return a shell file.
}
\description{
Farm is a computer cluster running slurm system.
Note, bigmem mem=8000 per cpu, hi med low mem=25000 per cpu, serial mem=1500 per cpu.
}
\examples{

}

