% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_jupyter.R
\name{set_jupyter}
\alias{set_jupyter}
\title{\code{Set up Jupyter Nodebook}}
\usage{
set_jupyter(port = 9999)
}
\arguments{
\item{port}{The port you are going to listen.}
}
\value{
nothing.
}
\description{
\code{Set up Jupyter Nodebook}
}
\examples{

## find ssh process ID and kill it in nodes
ps -ef | grep ssh | grep "jolyang"
kill -9 pid

## I don’t know what local port is available for me to use.
Linux: netstat -a | grep "999"
Mac OS X: lsof -i -P | grep LISTEN

set_jupyter(port=9998)

}

