% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_GenSel4.R
\name{run_GenSel4}
\alias{run_GenSel4}
\title{\code{Run GenSel on farm}}
\usage{
run_GenSel4(inputdf, inpdir = "largedata/", email = NULL,
  runinfo = c(FALSE, "bigmemh", 1))
}
\arguments{
\item{inputdf}{An input data.frame.}

\item{email}{Your email address that farm will email to once the jobs were done/failed.}

\item{runinfo}{Parameters specify the array job partition information.
A vector of c(FALSE, "bigmemh", "1"): 1) run or not, default=FALSE
2) -p partition name, default=bigmemh and 3) --cpus, default=1.
It will pass to \code{set_array_job}.}
}
\value{
return a batch of shell scripts.
}
\description{
GenSel4R
}
\details{
see more detail about gerpIBD
\url{https://github.com/yangjl/zmSNPtools}
}
\examples{
traits <- tolower(c("ASI", "DTP", "DTS", "EHT", "GY", "PHT", "TW"))
inputdf <- data.frame(pi=0.995,
   geno="/Users/yangjl/Documents/GWAS2_KRN/SNP/merged/geno_chr.newbin",
   trainpheno="/Users/yangjl/Documents/Heterosis_GWAS/pheno2011/reports/cd_GenSel_fullset.txt",
   testpheno="/Users/yangjl/Documents/Heterosis_GWAS/pheno2011/reports/cd_GenSel_fullset.txt",
   chainLength=1000, burnin=100, varGenotypic=1.4, varResidual=2)

run_GenSel4(inputdf, inpdir="slurm=script", email=NULL, runinfo = c(FALSE, "bigmemh", 1) )

}

