% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_arrayjob.R
\name{set_farm_job}
\alias{set_farm_job}
<<<<<<< HEAD
\title{Set up one farm job}
=======
\title{\code{Set up one farm job}}
>>>>>>> e2dfb9d60d5ce3dc09c5a85b3d3b2193373d0989
\usage{
set_farm_job(slurmsh = "largedata/GenSel/CL_test.sh",
  shcode = "sh largedata/myscript.sh", wd = NULL, jobid = "myjob",
  email = NULL)
}
\arguments{
<<<<<<< HEAD
\item{slurmsh}{Relative or absolute path and file name of your shell code, i.e. 'largedata/GenSel/CL_test.sh'.}

\item{shcode}{The commands inside your sh file.}

\item{wd}{Working directory, default=NULL => using your current directory.}

\item{jobid}{The job name show up in your sq "NAME' column.}
=======
\item{slurmsh}{Relative or absolute path and file name of your shell code, i.e. largedata/GenSel/CL_test.sh.}

\item{shcode}{The commands inside your sh file.}

\item{wd}{Working directory, default=NULL. It will use your current directory.}

\item{jobid}{The job name show up in your sq NAME column.}
>>>>>>> e2dfb9d60d5ce3dc09c5a85b3d3b2193373d0989

\item{email}{Your email address that farm will email to once the job was done/failed.}
}
\value{
a shell file.
}
\description{
<<<<<<< HEAD
'Farm' is a computer cluster running slurm system.
Note, 'bigmem' mem=8000/cpu, "hi/med/low' mem=25000/cpu, "serial" mem=1500/cpu.
}
\examples{


=======
Farm is a computer cluster running slurm system.
Note, bigmem mem=8000/cpu, hi/med/low mem=25000/cpu, serial mem=1500/cpu.
}
\examples{

>>>>>>> e2dfb9d60d5ce3dc09c5a85b3d3b2193373d0989
}

